+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.1.0
+
+- CachingSubCaseDS uses CachingSubCaseDSSampler to customize behaviour instead of function `decide_removal`.
+
+## 1.0.4
+
+- add attention pooling for `mmm.mtl_modules.shared_blocks.Grouper`
+
 ## 1.0.3
 
 - use default for local cache path if ML_DATA_CACHE is not set